Add dockerfiles for client & server applications, add scripts for running the application, Updated Readme with application initialization steps
