+++ overlay-network/docker-compose.yml
@@ -1,0 +1,36 @@
+services:
+  swarm-manager-comm:
+    env_file: ".env"
+    restart: always
+    container_name: swarm-manager-comm
+    image: alpine/socat
+    ports:
+      - "${EXTERNAL_IP}:2377:2377"
+    command: "TCP4-LISTEN:2377,fork TCP:${DOCKER_VM_IP}:2377"
+
+  swarm-node-comm-tcp:
+    env_file: ".env"
+    restart: always
+    container_name: swarm-node-comm-tcp
+    image: alpine/socat
+    ports:
+      - "${EXTERNAL_IP}:7946:7946/tcp"
+    command: "TCP4-LISTEN:7946,fork TCP4:${DOCKER_VM_IP}:7946"
+
+  swarm-node-comm-udp:
+    env_file: ".env"
+    restart: always
+    container_name: swarm-node-comm-udp
+    image: alpine/socat
+    ports:
+      - "${EXTERNAL_IP}:7946:7946/udp"
+    command: "UDP4-LISTEN:7946,fork UDP4:${DOCKER_VM_IP}:7946"
+
+  swarm-traffic-comm:
+    env_file: ".env"
+    restart: always
+    container_name: swarm-traffic-comm
+    image: alpine/socat
+    ports:
+      - "${EXTERNAL_IP}:4789:4789"
+    command: "UDP4-LISTEN:4789,fork UDP4:${DOCKER_VM_IP}:4789"
