+++ overlay-network/example.env
@@ -1,0 +1,2 @@
+EXTERNAL_IP=1.1.1.1
+DOCKER_VM_IP=172.17.0.1
