+++ server/Dockerfile
@@ -1,0 +1,9 @@
+FROM openjdk:11
+
+WORKDIR /app
+
+COPY . /app
+
+RUN ["javac", "BootstrapServer.java"]
+
+CMD ["java", "BootstrapServer"]
